DERBY-2386 Change {fn timestampdiff} to return a BIGINT. Contributed by Mayuresh Nirhali Mayuresh.Nirhali@Sun.COM
git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@525549 13f79535-47bb-0310-9956-ffa450edef68

+++ java/engine/org/apache/derby/impl/sql/compile/TernaryOperatorNode.java
@@ -823,7 +823,7 @@ public class TernaryOperatorNode extends ValueNode
 	{
         bindDateTimeArg( rightOperand, 2);
         bindDateTimeArg( receiver, 3);
-        setType(DataTypeDescriptor.getBuiltInDataTypeDescriptor( Types.INTEGER));
+        setType(DataTypeDescriptor.getBuiltInDataTypeDescriptor( Types.BIGINT));
         return this;
     } // End of timestampDiffBind
